feat: Refactor application with sorting algorithms, TUI improvements, and resource monitor enhancements
- Add 3 new sorting algorithms: Insertion Sort, Selection Sort, Bubble Sort
- Enhance resource monitoring with improved data tracking and CPU affinity
- Refactor main application with better algorithm selection and parameter handling
- Improve database manager error handling and query robustness
- Update CMakeLists.txt with new library targets and optimizations
- Complete TUI redesign:
  * Refactor index.ts with better state management
  * Improve execution screen with live metrics display
  * Enhanced settings screen with full field editing
  * Better visualizations and trend analysis
  * Fix settings persistence and auto-validation
  * Add sparklines for trend visualization with min/max values
- Update Makefile with simplified build targets
- Regenerate database with schema updates
- Update project summary documentation

This commit consolidates all recent improvements and prepares for v0.2 release.

+++ hpp/insertion_sort.hpp
@@ -6,20 +6,20 @@
 // Insertion sort function - O(n^2) time complexity
 void insertion_sort(std::vector<int>& arr)
 {
-    int n = arr.size();
-    for (int i = 1; i < n; ++i)
+    size_t n = arr.size();
+    for (size_t i = 1; i < n; ++i)
     {
         int key = arr[i];
-        int j = i - 1;
+        ssize_t j = i - 1; // Use signed type for downward loop
 
         // Move elements of arr[0..i-1] that are greater than key
         // to one position ahead of their current position
-        while (j >= 0 && arr[j] > key)
+        while (j >= 0 && arr[static_cast<size_t>(j)] > key)
         {
-            arr[j + 1] = arr[j];
+            arr[static_cast<size_t>(j) + 1] = arr[static_cast<size_t>(j)];
             --j;
         }
-        arr[j + 1] = key;
+        arr[static_cast<size_t>(j) + 1] = key;
     }
 }
